requirements.txt
streamlit
PyPDF2
Pillow
img2pdf
streamlit-sortables
